# swagger-gateway.yaml
swagger: "2.0"
info:
  title: Gateway API
  description: Piazza Core Services API
  version: "0.1.0"
  contact: 
    name: Patrick Doody
    email: patrick.doody@rgi-corp.com
    url: https://github.com/venicegeo/venice/wiki
host: pz-gateway.stage.geointservices.io
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /trigger:
    post:
      summary: Creates a Trigger
      description: "Creates a new Trigger with the Piazza Workflow component."
      parameters:
        - name: trigger
          in: body
          description: "The Trigger information to register. This defines the Conditions that must be hit in order for some Action to occur."
          required: true
          schema:
            $ref: '#/definitions/Trigger'
      tags:
        - Workflow
        - Trigger
      responses:
        200:
          description: "The ID of the newly created Trigger"
          schema:
            $ref: '#/definitions/TriggerResponse'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
    get:
      summary: List Triggers
      description: "Lists all of the defined Triggers in the Piazza Workflow component."
      parameters:
        - name: page
          in: query
          description: "Paginating large numbers of results. This will determine the starting page for the query."
          required: false
          type: number
        - name: per_page
          in: query
          description: "The number of results to be returned per query."
          required: false
          type: number
      tags:
        - Workflow
        - Trigger
      responses:
        200:
          description: "The list of Triggers."
          schema:
            $ref: '#/definitions/TriggerResults'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
  /trigger/{triggerId}:
    get:
      summary: Gets Metadata for a Trigger
      description: "Retrieves the Trigger definition for the Trigger matching the specified Trigger ID."
      parameters:
        - name: triggerId
          in: path
          description: "The ID of the Trigger to retrieve."
          required: true
          type: string
      tags:
        - Workflow
        - Trigger
      responses:
        200:
          description: "The Trigger metadata."
          schema:
            $ref: '#/definitions/Trigger'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      summary: Deletes a Trigger
      description: "Deletes a Trigger with the Workflow component. This Trigger will no longer listen for conditions for events to fire."
      parameters:
        - name: triggerId
          in: path
          description: "The ID of the Trigger to delete."
          required: true
          type: string
      tags:
        - Workflow
        - Trigger
      responses:
        200:
          description: "Confirmation of Trigger deletion.."
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
  /alert:
    get:
      summary: Get User Alerts
      description: "Gets all of the Alerts for the currently authenticated user. Alerts occur when a Trigger's conditions are met."
      parameters:
        - name: page
          in: query
          description: "Paginating large numbers of results. This will determine the starting page for the query."
          required: false
          type: number
        - name: per_page
          in: query
          description: "The number of results to be returned per query."
          required: false
          type: number
      tags:
        - Workflow
        - Alert
      responses:
        200:
          description: "The list of Alerts owned by the current User."
          schema:
            $ref: '#/definitions/AlertResults'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
  /alert/{alertId}:
    get:
      summary: Get Alert Information
      description: "Gets the metadata for a single Alert; referenced by ID."
      parameters:
        - name: alertId
          in: path
          description: "The ID of the Alert to retrieve metadata for."
          required: true
          type: string
      tags:
        - Workflow
        - Alert
      responses:
        200:
          description: "The Alert metadata."
          schema:
            $ref: '#/definitions/Alert'
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      summary: Delete Alert
      description: "Deletes an Alert; referenced by ID."
      parameters:
        - name: alertId
          in: path
          description: "The ID of the Alert to Delete."
          required: true
          type: string
      tags:
        - Workflow
        - Alert
      responses:
        200:
          description: "Confirmation of Delete."
        default:
          description: "Unexpected Error"
          schema:
            $ref: '#/definitions/ErrorResponse'
definitions:
  ErrorResponse:
    description: "The standard response object for an unexpected error within the Piazza core."
    type: object
    required:
      - message
      - origin
    properties:
      message:
        description: "The high-level description of the error message. May contain debug information useful for technical debugging."
        type: string
      origin:
        description: "The name of the component the error originated from."
        type: string
...
  TriggerResponse:
    type: object
    description: "A response specifying the ID of a newly created Trigger"
    required:
      - triggerId
    properties:
      triggerId:
        description: "The ID of the Trigger that has been created or updated from the request."
        type: string
...
  TriggerResults:
    type: object
    description: "Describes a list of returned query results of Trigger objects."
    properties:
      results:
        description: "The array of results."
        type: array
        items: 
          $ref: '#/definitions/Trigger'
      pagination:
        description: "The pagination metadata for this query."
        $ref: '#/definitions/Pagination'
  AlertResults:
    type: object
    description: "Describes a list of returned query results of Alert objects."
    properties:
      results:
        description: "The array of results."
        type: array
        items: 
          $ref: '#/definitions/Alert'
      pagination:
        description: "The pagination metadata for this query."
        $ref: '#/definitions/Pagination'
...
  Trigger:
    type: object
    description: "Defines an action that is to be taken when a specific Event occurs."
    properties:
      id:
        type: string
        description: "The unique identifier for this Event."
      title:
        type: string
        description: "The human-readable description of what this Trigger intends to accomplish."
      condition:
        type: object
        description: "Defines the trigger conditions, the query replacements, and the Job to execute upon this object becoming Triggered."
  Alert:
    type: object
    description: "Defines the event for when a Trigger has been fired based on a condition defined within that Trigger."
    properties:
      id: 
        type: string
        description: "The unique identifier for this Alert."
      trigger_id:
        type: string
        description: "The unique identifier for the Trigger that was hit."
      event_id:
        type: string
        description: "The unique identifier for the Event that triggered the Trigger."
  Pagination:
    type: object
    description: "Defines how to navigate large datasets in queries that contain more data than can be returned."
    properties:
      count:
        type: number
        description: "The number of results that matched the query."
      page:
        type: number
        description: "The current page returned."
      per_page:
        type: number
        description: "The current number of results returned per page."